requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==5.0.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
googlemaps==4.10.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.0.1
